+++ packages/sdk/src/errors/TransactionConfirmationError.ts
@@ -1,19 +1,27 @@
 import type { TransactionReceipt } from 'viem'
-import type { SvmReceipt, SuiReceipt, TronReceipt } from '../types/index.js'
+import type {
+  SvmReceipt,
+  TronReceipt,
+  LvmReceipt
+} from '../types/index.js'
 import type { TenderlyErrorInfo } from '../utils/getTenderlyDetails.js'
 
 export class TransactionConfirmationError extends Error {
   receipt:
     | TransactionReceipt
     | SvmReceipt
-    | SuiReceipt
     | TronReceipt
+    | LvmReceipt
     | undefined
   tenderlyError: TenderlyErrorInfo | undefined
 
   constructor(
     error: any,
-    receipt?: TransactionReceipt | SvmReceipt | SuiReceipt | TronReceipt,
+    receipt?:
+      | TransactionReceipt
+      | SvmReceipt
+      | TronReceipt
+      | LvmReceipt,
     tenderlyError?: TenderlyErrorInfo | null
   ) {
     super(error)