+++ packages/sdk/src/types/Execute.ts
@@ -1,4 +1,8 @@
-import type { SvmReceipt, SuiReceipt, TronReceipt } from '../types/index.js'
+import type {
+  SvmReceipt,
+  TronReceipt,
+  LvmReceipt
+} from '../types/index.js'
 import type { paths } from './api.js'
 import type { AxiosRequestConfig } from 'axios'
 import type { TransactionReceipt } from 'viem'
@@ -45,7 +49,11 @@ export type Execute = {
     depositAddress?: string
     items: {
       status: 'complete' | 'incomplete'
-      receipt?: TransactionReceipt | SvmReceipt | SuiReceipt | TronReceipt
+      receipt?:
+        | TransactionReceipt
+        | SvmReceipt
+        | TronReceipt
+        | LvmReceipt
       checkStatus?:
         | 'refund'
         | 'waiting'