+++ packages/sdk/src/utils/transaction.ts
@@ -6,8 +6,8 @@ import type {
   TransactionStepItem,
   paths,
   SvmReceipt,
-  SuiReceipt,
-  TronReceipt
+  TronReceipt,
+  LvmReceipt
 } from '../types/index.js'
 import { axios } from '../utils/axios.js'
 import type {
@@ -51,7 +51,11 @@ export async function sendTransactionSafely(
   isValidating?: (res?: AxiosResponse<any, any>) => void,
   details?: Execute['details'],
   setReceipt?: (
-    receipt: TransactionReceipt | SvmReceipt | SuiReceipt | TronReceipt
+    receipt:
+      | TransactionReceipt
+      | SvmReceipt
+      | TronReceipt
+      | LvmReceipt
   ) => void,
   setCheckStatus?: (
     checkStatus: NonNullable<Execute['steps'][0]['items']>[0]['checkStatus']
@@ -78,8 +82,8 @@ export async function sendTransactionSafely(
   let receipt:
     | TransactionReceipt
     | SvmReceipt
-    | SuiReceipt
     | TronReceipt
+    | LvmReceipt
     | undefined
   let transactionCancelled = false
   let confirmationError = false